js/xpconnect/tests/mochitest/test_getWebIDLCaller.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=968335
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 968335</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Cu.getCallerPrincipal (within JS-implemented WebIDL). **/

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({set: [['dom.expose_test_interfaces', true]]}, go);


  function go() {
    var t = new TestInterfaceJS();
    is(t.getCallerPrincipal(), location.origin,
       "Cu.getCallerPrincipal works right within JS-implemented WebIDL");

    try {
      SpecialPowers.Cu.getWebIDLCallerPrincipal();
      ok(false, "Should have thrown");
    } catch (e) {
      ok(/NOT_AVAILABLE/.test(SpecialPowers.wrap(e)),
         "API should throw when invoked outside of JS-implemented WebIDL");
    }

    SimpleTest.finish();
  }




  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=968335">Mozilla Bug 968335</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
